+++ src/definition-builder.ts
@@ -1,6 +1,7 @@
 import { DatumParameters, DefinitionConstr, DefinitionField } from './types';
 import { datumJsonToCbor } from 'lucid-cardano';
 import { DatumParameterKey } from './constants';
+import _ from 'lodash';
 
 export class DefinitionBuilder {
 
@@ -10,7 +11,11 @@ export class DefinitionBuilder {
      * Load a DEX definition file as a template for this builder.
      */
     public async loadDefinition(definition: DefinitionConstr): Promise<DefinitionBuilder> {
-        this._definition = structuredClone(definition);
+        this._definition = _.cloneDeepWith(definition, (value: any) => {
+            if (value instanceof Function) {
+                return value;
+            }
+        })
 
         return this;
     }
@@ -54,6 +59,10 @@ export class DefinitionBuilder {
      * Recursively set specified parameters.
      */
     private applyParameters(field: DefinitionField, mappedParameters: DatumParameters): DefinitionConstr {
+        if (field instanceof Function) {
+            return field(field, mappedParameters, false);
+        }
+
         if ('fields' in field) {
             if (typeof field.constructor === 'string') {
                 const parameterValue: any = mappedParameters[field.constructor as keyof typeof DatumParameterKey];
@@ -100,6 +109,20 @@ export class DefinitionBuilder {
      * Recursively pull parameters from datum using definition template.
      */
     private extractParameters(definedDefinition: DefinitionField, templateDefinition: DefinitionField, foundParameters: DatumParameters = {}): DatumParameters {
+        if (templateDefinition instanceof Function) {
+            templateDefinition(definedDefinition, foundParameters);
+
+            return foundParameters;
+        }
+
+        if (templateDefinition instanceof Array) {
+            templateDefinition.map((fieldParameter: DefinitionField, index: number) => {
+                return this.extractParameters(fieldParameter, templateDefinition[index], foundParameters);
+            }).forEach((parameters: DatumParameters) => {
+                foundParameters = {...foundParameters, ...parameters};
+            })
+        }
+
         if ('fields' in definedDefinition) {
             if (! ('fields' in templateDefinition)) {
                 throw new Error("Template definition does not match with 'fields'");