+++ src/providers/data/kupo-provider.ts
@@ -14,6 +14,7 @@ import {
 } from '@app/types';
 import axios, { AxiosInstance } from 'axios';
 import { Data } from 'lucid-cardano';
+import { appendSlash } from '@app/utils';
 
 export class KupoProvider extends BaseDataProvider {
 
@@ -35,7 +36,7 @@ export class KupoProvider extends BaseDataProvider {
 
         this._config = config;
         this._kupoApi = axios.create({
-            baseURL: this._requestConfig.proxyUrl + config.url,
+            baseURL: appendSlash(requestConfig.proxyUrl) + config.url,
             timeout: this._requestConfig.timeout,
             headers: {
                 'Content-Type': 'application/json',
@@ -73,7 +74,7 @@ export class KupoProvider extends BaseDataProvider {
     }
 
     public async assetTransactions(asset: Asset): Promise<Transaction[]> {
-        return this._kupoApi.get(`/matches/${asset.id('.')}`)
+        return this._kupoApi.get(`/matches/${asset.identifier('.')}`)
             .then((results: any) => {
                 return results.data.map((result: any) => {
                     return {
@@ -84,12 +85,12 @@ export class KupoProvider extends BaseDataProvider {
     }
 
     public async assetAddresses(asset: Asset): Promise<AssetAddress[]> {
-        return this._kupoApi.get(`/matches/${asset.id('.')}?unspent`)
+        return this._kupoApi.get(`/matches/${asset.identifier('.')}?unspent`)
             .then((results: any) => {
                 return results.data.map((result: any) => {
                     return {
                         address: result.address,
-                        quantity: BigInt(result.value.assets[asset.id('.')]),
+                        quantity: BigInt(result.value.assets[asset.identifier('.')]),
                     } as AssetAddress
                 }) as AssetAddress[];
             });
@@ -111,7 +112,7 @@ export class KupoProvider extends BaseDataProvider {
                     ];
                     Object.keys(utxo.value.assets).forEach((unit: string) => {
                         balances.push({
-                            asset: Asset.fromId(unit),
+                            asset: Asset.fromIdentifier(unit),
                             quantity: BigInt(utxo.value.assets[unit]),
                         });
                     });